Fix: empty alt attribute instead of "null" value when building img tag.
Omitting 'alt' attribute altogether indicates that the image is a key part of the content, and no textual equivalent is available. Setting this attribute to an empty string (alt="") indicates that this image is not a key part of the content, and that non-visual browsers may omit it from rendering. (see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img)